backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic-settings==2.6.0
python-multipart==0.0.12
httpx==0.27.2
email-validator==2.2.0
boto3==1.35.81
reportlab==4.2.5
alembic==1.14.0
pytest==8.3.3
pytest-cov==5.0.0
